Extract dependencies from coala-bears package.json
The dependencies were not constrained, which allowed
uncompatible versions of eslint and alex to be installed.
Using the coala-bears package.json ensures the Docker
image has the same tested versions as the CircleCI tests
run regularly for coala-bears.

Also this allows the Docker image to include other dependencies
added to coala-bears when new bears are added.