Reduce openSUSE packages installed and upgraded
Do not perform dist-upgrade or install pattern devel_basis.
They are separate commands each needing to load the package
database and add an extra 200Mb to the installed image.

Dist-upgrade updates 44 packages, most of which are not used
(e.g. krb5-mini) or the existing version is ok (e.g. bash).
The only components needing upgrading are libgcc_s1 and libstdc++6.

Of the devel_basis, only `patch` is needed for ruby gem `pg_query`,
which is used by `sqlint` for `SQLintBear`.